refactor: 각종 에러 수정

+++ src/main/java/com/example/book_your_seat/concert/controller/dto/ConcertResponse.java
@@ -17,7 +17,7 @@ public final class ConcertResponse {
     private final Integer totalStock;
     private final LocalDate startDate;
     private final LocalDate endDate;
-    private final Integer price;
+    private final Long price;
     private final Integer startHour;
     private final LocalDateTime reservationStartAt;
     private final List<Review> reviews;
